Moved team discriminators into Team and dropped rest
