+++ server/models/team/team.js
@@ -56,6 +56,81 @@ function validateNational(national) {
   return Joi.validate(national, schema, { "allowUnknown": true });
 };
 
+const Alien = Team.discriminator("Alien", new Schema({
+  type: { type: String, default: 'Alien' },
+  roles: [RoleSchema],
+  actionPts: { type: Number, default: 25 },
+  agents: { type: Number, min: 0, default: 0 },
+  sciRate: { type: Number, default: 25 }
+}));
+
+function validateAlien(alien) {
+  //modelDebugger(`Validating ${alien.name}...`);
+
+  const schema = {
+      teamCode: Joi.string().min(2).max(3).required().uppercase(),
+      name: Joi.string().min(2).max(50).required(),
+      shortName: Joi.string().min(2).max(30),
+      teamType: Joi.string().min(1).max(1).uppercase()
+    };
+  
+  return Joi.validate(alien, schema, { "allowUnknown": true });
+};
+
+const Control = Team.discriminator("Control", new Schema({
+  type: { type: String, default: 'Control' },
+  sciRate: { type: Number, default: 25},
+  roles: [RoleSchema]
+}));
+
+function validateControl(control) {
+  //modelDebugger(`Validating ${control.name}...`);
+
+  const schema = {
+      teamCode: Joi.string().min(2).max(3).required().uppercase(),
+      name: Joi.string().min(2).max(50).required(),
+      shortName: Joi.string().min(2).max(30),
+      teamType: Joi.string().min(1).max(1).uppercase()
+    };
+  
+  return Joi.validate(control, schema, { "allowUnknown": true });
+};
+
+const Media = Team.discriminator("Media", new Schema({
+  type: { type: String, default: 'Media' },
+  agents: { type: Number }
+}));
+
+function validateMedia(media) {
+  //modelDebugger(`Validating ${media.name}...`);
+
+  const schema = {
+      teamCode: Joi.string().min(2).max(3).required().uppercase(),
+      name: Joi.string().min(2).max(50).required(),
+      shortName: Joi.string().min(2).max(30),
+      teamType: Joi.string().min(1).max(1).uppercase()
+    };
+  
+  return Joi.validate(media, schema, { "allowUnknown": true });
+};
+
+const Npc = Team.discriminator("Npc", new Schema({
+  type: { type: String, default: 'Npc' }
+}));
+
+function validateNpc(npc) {
+  //modelDebugger(`Validating ${npc.name}...`);
+
+  const schema = {
+      teamCode: Joi.string().min(2).max(3).required().uppercase(),
+      name: Joi.string().min(2).max(50).required(),
+      shortName: Joi.string().min(2).max(30),
+      teamType: Joi.string().min(1).max(1).uppercase()
+    };
+  
+  return Joi.validate(npc, schema, { "allowUnknown": true });
+};
+
 function validateTeam(team) {
   //modelDebugger(`Validating ${team.name}...`);
 
@@ -95,4 +170,5 @@ async function getSciRate(team_id) {
   }
 };
 
-module.exports = { Team, validateTeam, getTeam, getPR, getSciRate, RoleSchema }
+module.exports = { Team, validateTeam, getTeam, getPR, getSciRate, National, validateNational, Alien, validateAlien,
+                   Control, validateControl, Media, validateMedia, Npc, validateNpc }
