+++ server/util/dataInit/initRefLoad.js
@@ -21,12 +21,7 @@ const bodyParser = require('body-parser');
 // Country Model - Using Mongoose Model
 const { Zone, validateZone } = require('../../models/zone');
 const { Country, validateCountry } = require('../../models/country'); 
-const { Team, validateTeam } = require('../../models/team/team');
-const { Alien, validateAlien } = require('../../models/team/alien');
-const { Control, validateControl } = require('../../models/team/control');
-const { Media, validateMedia } = require('../../models/team/media');
-const { National, validateNational } = require('../../models/team/national');
-const { Npc, validateNpc } = require('../../models/team/npc');
+const { Team, validateTeam, National, validateNational, Alien, validateAlien, Control, validateControl, Npc, validateNpc, Media, validateMedia } = require('../../models/team/team');
 const Research = require('../../models/sci/research');
 const Log = require('../../models/logs/log');
